Fix issue with analysing single signal in bispectrum analysis
MODA threw an error when a single signal was loaded in bispectrum analysis; this was due to a previous change where only an even number of signals was allowed in windows which use signal pairs. Now a single signal is allowed, but an odd number of signals greater than 2 is cut down to its closest even number (e.g. the last signal will be removed when 5 signals are loaded).