Show error dialog when handles.fc does not exist
This error is related to the previous issue where an error occurred when clicking "Add marked region" after performing a transform using a min/max frequency.

For unknown reasons, after a transform is performed with a specified fmin/fmax, some fields are deleted from 'handles' (for example, handles.fc and handles.freqarr).

This causes myriad problems including an error when 'handles.fc' does not exist, which is the problem addressed by this commit. As a temporary solution, an error dialog is now shown which prompts the user to re-calculate the transform.